feat: ceiling, floor RGB값을 4byte 값의 데이터로 저장#26
