+++ Makefile
@@ -6,7 +6,7 @@
 #    By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/04/27 15:13:41 by kipark            #+#    #+#              #
-#    Updated: 2022/11/18 18:25:23 by jayoon           ###   ########.fr        #
+#    Updated: 2022/11/18 20:54:30 by jayoon           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -65,7 +65,8 @@ SRCS_PARSE_MAP				:=	init_map_info.c			\
 								is_valid_character.c	\
 								get_cnt.c				\
 								is_start_position.c		\
-								get_start_pos_info.c
+								get_start_pos_info.c	\
+								get_color.c
 SRCS_PARSE_MAP				:=	$(addprefix $(DIR_PARSE_MAP)/, $(SRCS_PARSE_MAP))
 
 SRCS_EVENT					:=	key.c
