+++ srcs/parse_map/process_color.c
@@ -6,7 +6,7 @@
 /*   By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 13:14:07 by jayoon            #+#    #+#             */
-/*   Updated: 2022/10/29 18:11:10 by jayoon           ###   ########.fr       */
+/*   Updated: 2022/11/18 20:51:57 by jayoon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,37 +35,6 @@ static size_t	count_comma(char *str)
 		++str;
 	}
 	return (cnt);
-}
-
-static void	check_0_to_255(int num)
-{
-	if (num < 0 || num > 255)
-		print_error_str(MSG_ERR_ELEMENT);
-}
-
-static void	get_color(t_map_info *map_info, t_identifier num_iden, char *str)
-{
-	char			**arr_rgb;
-	t_color			color;
-	int				num;
-
-	color = RED;
-	arr_rgb = ft_split(str, ',');
-	while (arr_rgb[color])
-	{
-		num = cub3d_atoi(arr_rgb[color]);
-		if (num == -1)
-			print_error_str(MSG_ERR_ELEMENT);
-		check_0_to_255(num);
-		if (num_iden == FLOOR)
-			map_info->floor[color] = num;
-		else
-			map_info->ceiling[color] = num;
-		++color;
-	}
-	if (color != 3)
-		print_error_str(MSG_ERR_ELEMENT);
-	ft_safe_free_two_dimentions_arr(arr_rgb);
 }
 
 /*
